+++ babel.config.js
@@ -6,6 +6,6 @@ module.exports = {
     ],
      "plugins": [
       ["@babel/plugin-proposal-decorators", { "legacy": true }],
-     /*  ["@babel/plugin-proposal-class-properties", { "loose": true}] */
+      ["@babel/plugin-proposal-class-properties", { "loose": true}]
   ] 
   };
