+++ jest.config.ts
@@ -1,9 +1,11 @@
+
 /*
  * For a detailed explanation regarding each configuration property and type check, visit:
  * https://jestjs.io/docs/configuration
  */
 
 export default {
+
   // All imported modules in your tests should be mocked automatically
   // automock: false,
 
@@ -88,7 +90,9 @@ export default {
   // ],
 
   // A map from regular expressions to module names or to arrays of module names that allow to stub out resources with a single module
-  // moduleNameMapper: {},
+  /*  moduleNameMapper:pathsToModuleNameMapper(compilerOptions.paths,{
+     prefix:"<rootDir/src/>",
+   }) */
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
   // modulePathIgnorePatterns: [],
