+++ package.json
@@ -4,8 +4,8 @@
   "main": "index.js",
   "license": "MIT",
   "scripts": {
-    "dev": "ts-node-dev --inspect --transpile-only --ignore-watch node_modules --respawn src/server.ts",
-    "typeorm": "ts-node-dev ./node_modules/typeorm/cli",
+    "dev": "ts-node-dev -r tsconfig-paths/register --inspect --transpile-only --ignore-watch node_modules --respawn src/shared/infra/http/server.ts",
+    "typeorm": "ts-node-dev -r tsconfig-paths/register ./node_modules/typeorm/cli",
     "test": "jest"
   },
   "devDependencies": {
@@ -32,6 +32,7 @@
     "prettier": "^2.0.4",
     "ts-jest": "<23.10.0",
     "ts-node-dev": "^1.0.0-pre.44",
+    "tsconfig-paths": "^4.0.0",
     "typescript": "^3.8.3"
   },
   "dependencies": {
