+++ src/modules/accounts/useCases/authenticationUser/AuthenticationUserCase.ts
@@ -2,8 +2,8 @@ import {compare} from "bcryptjs";
 import {sign} from "jsonwebtoken";
 import "reflect-metadata"
 import { inject, injectable } from "tsyringe";
-import { AppError } from "../../../../errors/AppError";
-import { IUserRepository } from "../../repository/User/IUserRepository";
+import { AppError } from "@errors/AppError";
+import { IUserRepository } from "@modules/accounts/repository/User/IUserRepository";
 
 interface IRequest {
   email: string;
