+++ src/modules/cars/infra/typeorm/entity/Categoria.ts
