+++ src/modules/cars/infra/typeorm/entity/Categoria_test_.ts
@@ -1,5 +1,5 @@
 import "reflect-metadata"
-import { Column, CreateDateColumn, Entity, PrimaryColumn } from "typeorm";
+import { Entity } from "typeorm";
 import { v4 as uuidv4 } from "uuid";
 @Entity("categories")
 export class Categoria_test_ {
