+++ src/modules/cars/infra/typeorm/entity/Specification.ts
