+++ src/modules/cars/repositories/Categotery/dtos/ICreateCategoryDTO.ts
@@ -1,0 +1,5 @@
+interface ICreateCategariaDTO {
+  name: string;
+  discricao: string;
+}
+export {ICreateCategariaDTO}
