+++ src/modules/cars/repositories/Specification/dtos/ICreateSpecificationDTO.ts
@@ -1,0 +1,5 @@
+interface ICreateSpecificationDTO {
+    name: string;
+    discricao: string;
+  }
+  export {ICreateSpecificationDTO}
