+++ src/modules/cars/useCases/Specifications/CreateSpecification/CreateSpecificationUseCase.ts
@@ -1,6 +1,9 @@
-import { ISpecificationsRepository } from "../../../repositories/Specification/ISpecificationsRepository";
+
 import { inject,injectable} from 'tsyringe'
-import { AppError } from "../../../../../errors/AppError";
+import { AppError } from "@errors/AppError";
+import { ISpecificationsRepository } from '@modules/cars/repositories/Specification/ISpecificationsRepository';
+
+
 interface IRequest {
   name: string;
   discricao: string;
