+++ src/shared/container/index.ts
@@ -1,26 +1,23 @@
-import "reflect-metadata"
+import "reflect-metadata";
 import { container } from "tsyringe";
-import { UsersRepository } from "../../modules/accounts/repository/User/implementations/UsersRepository";
-import { IUserRepository } from "../../modules/accounts/repository/User/IUserRepository";
-import { ICategoryRepository } from "../../modules/cars/repositories/Category/ICategoryRepository";
-import { CategoriasRepository } from "../../modules/cars/repositories/Category/implementations/CategoriasRepository";
-import { CategoryRepositoryInMemory } from "../../modules/cars/repositories/Category/in-memory/CategoryRepositoryInMemory";
-import { SpecificationsRepository } from "../../modules/cars/repositories/Specification/implementations/SpecificationsRepository";
-import { ISpecificationsRepository } from "../../modules/cars/repositories/Specification/ISpecificationsRepository";
+import { UsersRepository } from "@modules/accounts/infra/typeorm/repository/UsersRepository";
+import { IUserRepository } from "@modules/accounts/repository/User/IUserRepository";
 
-// Interface ICategoryRepository
+import { CategoriasRepository } from "@modules/cars/infra/typeorm/repositories/Category/CategoriasRepository";
+import { SpecificationsRepository } from "@modules/cars/infra/typeorm/repositories/Specification/SpecificationsRepository";
+
+import { ICategoryRepository } from "@modules/cars/repositories/Categotery/ICategoryRepository";
+import { ISpecificationsRepository } from "@modules/cars/repositories/Specification/ISpecificationsRepository";
+
 container.registerSingleton<ICategoryRepository>(
-    "CategoriasRepository",
+  "CategoriasRepository",
   CategoriasRepository
-); 
-
-
+);
 container.registerSingleton<ISpecificationsRepository>(
-    "SpecificationsRepository",
-    SpecificationsRepository
+  "SpecificationsRepository",
+  SpecificationsRepository
 );
 container.registerSingleton<IUserRepository>(
   "UsersRepository",
   UsersRepository
 );
-
