+++ src/shared/infra/http/routes/specification.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
-import { ensureAuthenticated } from "../middlewares/ensureAuthenticated";
-import { CreateSpecificationController } from "../modules/cars/useCases/Specifications/CreateSpecification/CreateSpecificationController";
-import { ListSpecificationController } from "../modules/cars/useCases/Specifications/ListSpecification/ListSpecificationController";
+import { CreateSpecificationController } from "@modules/cars/useCases/Specifications/CreateSpecification/CreateSpecificationController";
+import { ListSpecificationController } from "@modules/cars/useCases/Specifications/ListSpecification/ListSpecificationController";
+import { ensureAuthenticated } from "@shared/infra/http/middlewares/ensureAuthenticated";
 
 const specificationRoutes = Router();
