+++ src/shared/infra/http/server.ts
@@ -1,13 +1,14 @@
 import express, { json, NextFunction, Response, Request } from "express";
 import "express-async-errors";
-import "./database";
+import "../typeorm";
 import { routers } from "./routes";
 import swaggerUri from "swagger-ui-express";
-import swaggerFile from "./swagger.json";
-import swaggerFilev2 from "./swagger.v2.json";
-import "./shared/container";
-import { AppError } from "./errors/AppError";
+import swaggerFile from "../../../swagger.json";
+
+import "@shared/container";
+
 import dotenv from "dotenv";
+import { AppError } from "@errors/AppError";
 dotenv.config();
 
 const app = express();
@@ -15,7 +16,6 @@ app.use(express.json());
 app.use(routers);
 
 app.use("/api-docs", swaggerUri.serve, swaggerUri.setup(swaggerFile));
-app.use("/api-v2.docs", swaggerUri.serve, swaggerUri.setup(swaggerFilev2));
 ///////////////////////////////////////////
 app.use((err: Error, req: Request, resp: Response, next: NextFunction) => {
   if (err instanceof AppError) {
