--- src/swagger.v2.json
@@ -1,79 +1,0 @@
-{
-    "openapi": "3.0.0",
-    "info": {
-        "title": "RENTACARS Docs",
-        "description": "Esta api é sobre aluguer de viaturas",
-        "version": "2.0.0",
-        "contact": {
-            "email": "mechackantonio@gmail.com"  
-        }
-    },
-    "paths":{
-        "/categaria":{
-            "post":{
-                "tags": ["Categoria"],
-                "sumary": "Criação de categoria",
-                "description":"criação de uma nova categoria",
-                "requestBody":{
-                    "content":{
-                        "application/json":{
-                            "schema":{
-                                "type": "object",
-                                "properties":{
-                                    "name": {
-                                        "type": "string"
-                                        
-                                    },
-                                    "description": {
-                                        "type": "string"
-                                    }
-                                },
-                                "example":{
-                                    "name": "Nome da categoria",
-                                    "description": "descrição da categoria"
-                                }
-                            }
-                        }
-                    }
-                },
-                "responses":{
-                    "201":{
-                        "description": "Created"
-                    },
-                    "500": {
-                        "description": "Categoria já existe"
-                    }
-                }
-            },
-            "get":{
-                "tags": ["Categoria"],
-                "sumary": "Listar todas categoria",
-                "description": "Listar todas categorias",
-                "responses":{
-                    "200":{
-                       "description": "Sucesso",
-                       "content":{
-                           "application/json": {
-                               "schema":{
-                                   "type": "array",
-                                   "items":{
-                                       "type":"object",
-                                       "properties":{
-                                           "name":{
-                                               "type":"string"
-                                           },
-                                           "description":{
-                                               "type":"string"
-                                           }
-                                       }
-                                   }
-                               }
-                           }
-                       } 
-                    }
-                }
-            }
-           
-        }
-    }
-}
